requirements.txt
torch>=2.0
torchvision>=0.15

numpy>=1.23
opencv-python>=4.8
Pillow>=9.5

matplotlib>=3.7
tqdm>=4.65

scikit-learn>=1.2

jupyter